Fix send batch typing.
