+++ modules/sqs/src/main/scala/vinyldns/sqs/queue/SqsMessageQueue.scala
@@ -16,6 +16,7 @@
 
 package vinyldns.sqs.queue
 
+import java.util.Base64
 import java.util.concurrent.TimeUnit.SECONDS
 
 import cats.data.NonEmptyList
@@ -27,11 +28,12 @@ import com.amazonaws.services.sqs.model._
 import com.amazonaws.services.sqs.{AmazonSQSAsync, AmazonSQSAsyncClientBuilder}
 import com.amazonaws.{AmazonWebServiceRequest, AmazonWebServiceResult}
 import com.typesafe.config.{Config, ConfigFactory}
-import vinyldns.core.domain.ZoneCommand
+import org.slf4j.LoggerFactory
+import vinyldns.core.domain.{RecordSetChange, ZoneChange, ZoneCommand}
 import vinyldns.core.protobuf.ProtobufConversions
 import vinyldns.core.queue._
 import vinyldns.core.route.Monitor
-import vinyldns.sqs.queue.SqsConverters._
+import vinyldns.proto.VinylDNSProto
 
 import scala.collection.JavaConverters._
 import scala.concurrent.duration.FiniteDuration
@@ -43,6 +45,8 @@ final case class SqsMessageHandle(receiptHandle: String) extends MessageHandle
 case class SqsMessageQueue(queueUrl: String, client: AmazonSQSAsync)
     extends MessageQueue
     with ProtobufConversions {
+
+  import SqsMessageQueue._
 
   // Helper function for monitoring instrumentation
   private def monitored[A](name: String)(f: => IO[A]): IO[A] = {
@@ -110,16 +114,22 @@ case class SqsMessageQueue(queueUrl: String, client: AmazonSQSAsync)
           .withQueueUrl(queueUrl),
         client.sendMessageAsync)).map(_ => ())
 
-  def send[A <: ZoneCommand](messages: NonEmptyList[A]): IO[SendBatchResult] =
+  def send[A <: ZoneCommand](cmds: NonEmptyList[A]): IO[SendBatchResult] =
     monitored("sqs.sendMessageBatch")(
       sqsAsync[SendMessageBatchRequest, SendMessageBatchResult](
-        toSendMessageRequest(messages)
+        toSendMessageBatchRequest(cmds)
           .withQueueUrl(queueUrl),
         client.sendMessageBatchAsync))
       .map { batchResult =>
-        val idLookupMap = messages.toList.map(zc => zc.id -> zc).toMap
-        val successes = batchResult.getSuccessful.asScala.map(fromMessage(_, idLookupMap)).toList
-        val failures = batchResult.getFailed.asScala.map(fromMessage(_, idLookupMap)).toList
+        val successfulIds = batchResult.getSuccessful.asScala.map(_.getId)
+        val successes = cmds.toList.filter(cmd => successfulIds.contains(cmd.id))
+
+        val failureIds = batchResult.getFailed.asScala.map(_.getId)
+        val failureMessages = batchResult.getFailed.asScala.map(_.getMessage)
+        val failures =
+          cmds.toList.filter(cmd => failureIds.contains(cmd.id)).zip(failureMessages).map {
+            case (cmd, msg) => (new Exception(msg), cmd)
+          }
         SendBatchResult(successes, failures)
       }
 
@@ -134,7 +144,39 @@ case class SqsMessageQueue(queueUrl: String, client: AmazonSQSAsync)
       )).map(_ => ())
 }
 
-object SqsMessageQueue {
+object SqsMessageQueue extends ProtobufConversions {
+  private val logger = LoggerFactory.getLogger("vinyldns.sqs.queue.SqsConverters")
+
+  sealed abstract class SqsMessageType(val name: String)
+  case object SqsRecordSetChangeMessage extends SqsMessageType("SqsRecordSetChangeMessage")
+  case object SqsZoneChangeMessage extends SqsMessageType("SqsZoneChangeMessage")
+
+  implicit class AmazonWebServiceRequestImprovements[A <: AmazonWebServiceRequest](
+      baseMessage: SendMessageRequest) {
+    def withMessageType(messageTypeName: String): SendMessageRequest =
+      baseMessage
+        .withMessageAttributes(
+          Map(
+            "message-type" -> new MessageAttributeValue()
+              .withStringValue(messageTypeName)
+              .withDataType("String")
+          ).asJava
+        )
+  }
+
+  implicit class SendMessageBatchRequestEntryImprovements(
+      baseRequestEntry: SendMessageBatchRequestEntry) {
+    def withMessageType(messageTypeName: String): SendMessageBatchRequestEntry =
+      baseRequestEntry
+        .withMessageAttributes(
+          Map(
+            "message-type" -> new MessageAttributeValue()
+              .withStringValue(messageTypeName)
+              .withDataType("String")
+          ).asJava
+        )
+  }
+
   def apply(config: Config = ConfigFactory.load().getConfig("sqs")): SqsMessageQueue = {
     val accessKey = config.getString("access-key")
     val secretKey = config.getString("secret-key")
@@ -151,5 +193,67 @@ object SqsMessageQueue {
         .build()
 
     new SqsMessageQueue(queueUrl, client)
+  }
+
+  // Helper function to serialize message body
+  def messageData[A <: ZoneCommand](cmd: A): String = cmd match {
+    case rsc: RecordSetChange => Base64.getEncoder.encodeToString(toPB(rsc).toByteArray)
+    case zc: ZoneChange => Base64.getEncoder.encodeToString(toPB(zc).toByteArray)
+  }
+
+  // Helper function to generate message type attribute
+  def messageType[A <: ZoneCommand](cmd: A): String = cmd match {
+    case _: RecordSetChange => SqsRecordSetChangeMessage.name
+    case _: ZoneChange => SqsZoneChangeMessage.name
+  }
+
+  def parseMessageType(messageType: String): SqsMessageType = messageType match {
+    case SqsRecordSetChangeMessage.name => SqsRecordSetChangeMessage
+    case SqsZoneChangeMessage.name => SqsZoneChangeMessage
+  }
+
+  def toSendMessageRequest(zoneCommand: ZoneCommand): SendMessageRequest = {
+    val messageTypeBytesTuple = zoneCommand match {
+      case rsc: RecordSetChange => (SqsRecordSetChangeMessage.name, toPB(rsc).toByteArray)
+      case zc: ZoneChange => (SqsZoneChangeMessage.name, toPB(zc).toByteArray)
+    }
+
+    messageTypeBytesTuple match {
+      case (messageType, messageBytes) =>
+        new SendMessageRequest()
+          .withMessageBody(Base64.getEncoder.encodeToString(messageBytes))
+          .withMessageType(messageType)
+    }
+  }
+
+  def toSendMessageBatchRequest[A <: ZoneCommand](
+      commands: NonEmptyList[A]): SendMessageBatchRequest = {
+    // convert each message into an entry
+    val entries = commands
+      .map(cmd => (cmd, messageData(cmd)))
+      .map {
+        case (cmd, msgBody) =>
+          new SendMessageBatchRequestEntry()
+            .withMessageBody(msgBody)
+            .withId(cmd.id)
+            .withMessageType(messageType(cmd))
+      }
+      .toList
+      .asJava
+
+    new SendMessageBatchRequest()
+      .withEntries(entries)
+  }
+
+  def fromMessage(message: Message): ZoneCommand = {
+    logger.info(
+      s"Received message with attributes ${message.getMessageAttributes.asScala}, ${message.getAttributes.asScala}")
+    val messageType = message.getMessageAttributes.asScala("message-type").getStringValue
+    val messageBytes = Base64.getDecoder.decode(message.getBody)
+    parseMessageType(messageType) match {
+      case SqsRecordSetChangeMessage =>
+        fromPB(VinylDNSProto.RecordSetChange.parseFrom(messageBytes))
+      case SqsZoneChangeMessage => fromPB(VinylDNSProto.ZoneChange.parseFrom(messageBytes))
+    }
   }
 }
